requirements.txt
# Core Trading & Market Data
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning (Optional - for ML signal filtering)
scikit-learn>=1.3.0
xgboost>=2.0.0

# Notifications (Optional - for Telegram alerts)
python-telegram-bot>=20.4

# Configuration & Utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Development & Testing (Optional)
pytest>=7.4.0
